requirements.txt
asgiref==3.8.1
black==24.10.0
boto3==1.35.52
botocore==1.35.52
click==8.1.7
dj-database-url==0.5.0
Django==4.2
django-allauth==0.51.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
Markdown==3.7
oauthlib==3.2.2
pathspec==0.12.1
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
s3transfer==0.10.3
sqlparse==0.5.1
stripe==11.1.1
